unsure about if two nested loops is enough to test loop - it only ever uses one RAM cell i think, for x, with y held in the scratch register. so that isn't testing movement of the stack pointer much.
as it is now, it seems to lose the outer counter

but put in some delay states after setting the read data address, and now it works - so I guess the RAM wasn't responding fast enough.

maybe i only need 1 delay state here, instead of two.

and is there a principled approach here? eg that it takes the RAM register one clock cycle to respond after the address, so then another clock cycle to see the result?